Move QueryHandler annotation to annotation package
Move QueryHandler annotation to annotation package. This allows us to
introduce a QueryHandler interface in the base command handling package
. Furthermore, this shift serves as a precursor to all annotation logic
being moved to a separate module

#3067
