+++ messaging/src/main/java/org/axonframework/util/HandlerTypeResolver.java
@@ -18,7 +18,7 @@ package org.axonframework.util;
 
 import org.axonframework.commandhandling.annotation.CommandHandler;
 import org.axonframework.eventhandling.annotation.EventHandler;
-import org.axonframework.queryhandling.QueryHandler;
+import org.axonframework.queryhandling.annotation.QueryHandler;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
