+++ messaging/src/test/java/org/axonframework/queryhandling/annotation/AnnotationQueryHandlerAdapterTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2010-2024. Axon Framework
+ * Copyright (c) 2010-2025. Axon Framework
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,7 +25,6 @@ import org.axonframework.messaging.interceptors.MessageHandlerInterceptor;
 import org.axonframework.queryhandling.GenericQueryMessage;
 import org.axonframework.queryhandling.NoHandlerForQueryException;
 import org.axonframework.queryhandling.QueryBus;
-import org.axonframework.queryhandling.QueryHandler;
 import org.axonframework.queryhandling.QueryMessage;
 import org.axonframework.utils.MockException;
 import org.junit.jupiter.api.*;
